More UWP prep
