+++ Common/FileUtil.cpp
@@ -556,11 +556,12 @@ bool CreateEmptyFile(const std::string &filename)
 	return true;
 }
 
-#if !PPSSPP_PLATFORM(UWP)
-
 // Deletes the given directory and anything under it. Returns true on success.
 bool DeleteDirRecursively(const std::string &directory)
 {
+#if PPSSPP_PLATFORM(UWP)
+	return false;
+#else
 	INFO_LOG(COMMON, "DeleteDirRecursively: %s", directory.c_str());
 
 #ifdef _WIN32
@@ -629,11 +630,10 @@ bool DeleteDirRecursively(const std::string &directory)
 	closedir(dirp);
 #endif
 	File::DeleteDir(directory);
-		
 	return true;
+#endif
 }
 
-#endif
 
 // Create directory and copy contents (does not overwrite existing files)
 void CopyDir(const std::string &source_path, const std::string &dest_path)
