+++ GPU/Common/ShaderTranslation.cpp
@@ -15,7 +15,7 @@
 // Official git repository and contact information can be found at
 // https://github.com/hrydgard/ppsspp and http://www.ppsspp.org/.
 
-#ifndef ANDROID
+#if !defined(ANDROID)
 
 #include <memory>
 #include <vector>
