+++ UI/NativeApp.cpp
@@ -634,7 +634,11 @@ void NativeInitGraphics(GraphicsContext *graphicsContext) {
 
 #ifdef _WIN32
 	winAudioBackend = CreateAudioBackend((AudioBackendType)g_Config.iAudioBackend);
+#if PPSSPP_PLATFORM(UWP)
+	// TODO UWP
+#else
 	winAudioBackend->Init(MainWindow::GetHWND(), &Win32Mix, 44100);
+#endif
 #endif
 
 	g_gameInfoCache = new GameInfoCache();
