+++ ext/native/ext/jpge/jpge.cpp
@@ -896,83 +896,6 @@ bool jpeg_encoder::process_scanline(const void* pScanline)
   return m_all_stream_writes_succeeded;
 }
 
-class cfile_stream : public output_stream
-{
-   cfile_stream(const cfile_stream &);
-   cfile_stream &operator= (const cfile_stream &);
-
-   FILE* m_pFile;
-   bool m_bStatus;
-
-public:
-   cfile_stream() : m_pFile(NULL), m_bStatus(false) { }
-
-   virtual ~cfile_stream()
-   {
-      close();
-   }
-
-   bool open(const char *pFilename)
-   {
-      close();
-      m_pFile = fopen(pFilename, "wb");
-      m_bStatus = (m_pFile != NULL);
-      return m_bStatus;
-   }
-
-   bool close()
-   {
-      if (m_pFile)
-      {
-         if (fclose(m_pFile) == EOF)
-         {
-            m_bStatus = false;
-         }
-         m_pFile = NULL;
-      }
-      return m_bStatus;
-   }
-
-   virtual bool put_buf(const void* pBuf, int len)
-   {
-      m_bStatus = m_bStatus && (fwrite(pBuf, len, 1, m_pFile) == 1);
-      return m_bStatus;
-   }
-
-   uint get_size() const
-   {
-      return m_pFile ? ftell(m_pFile) : 0;
-   }
-};
-
-// Writes JPEG image to file.
-bool compress_image_to_jpeg_file(const char *pFilename, int width, int height, int num_channels, const uint8 *pImage_data, const params &comp_params)
-{
-  cfile_stream dst_stream;
-  if (!dst_stream.open(pFilename))
-    return false;
-
-  jpge::jpeg_encoder dst_image;
-  if (!dst_image.init(&dst_stream, width, height, num_channels, comp_params))
-    return false;
-
-  for (uint pass_index = 0; pass_index < dst_image.get_total_passes(); pass_index++)
-  {
-    for (int i = 0; i < height; i++)
-    {
-       const uint8* pBuf = pImage_data + i * width * num_channels;
-       if (!dst_image.process_scanline(pBuf))
-          return false;
-    }
-    if (!dst_image.process_scanline(NULL))
-       return false;
-  }
-
-  dst_image.deinit();
-
-  return dst_stream.close();
-}
-
 class memory_stream : public output_stream
 {
    memory_stream(const memory_stream &);
