+++ ext/native/net/http_server.cpp
@@ -2,7 +2,10 @@
 
 #ifdef _WIN32
 
+#ifndef NOMINMAX
 #define NOMINMAX
+#endif
+
 #include <winsock2.h>
 #include <ws2tcpip.h>
 #include <io.h>
