+++ ext/native/net/sinks.cpp
@@ -1,6 +1,8 @@
 #ifdef _WIN32
 
+#ifndef NOMINMAX
 #define NOMINMAX
+#endif
 #include <winsock2.h>
 #include <ws2tcpip.h>
 #include <io.h>
