+++ ext/udis86/udis86.c
@@ -168,7 +168,7 @@ ud_insn_hex(struct ud* u)
     /* for each byte used to decode instruction */
     for (i = 0; i < ud_insn_len(u) && i < sizeof(u->insn_hexcode) / 2;
          ++i, ++src_ptr) {
-      sprintf(src_hex, "%02x", *src_ptr & 0xFF);
+      snprintf(src_hex, 64 - i * 2, "%02x", *src_ptr & 0xFF);
       src_hex += 2;
     }
   }
