+++ ppsspp_config.h
@@ -77,7 +77,7 @@
     // Covers both 32 and 64bit Windows
     #define PPSSPP_PLATFORM_WINDOWS 1
     // UWP trickery
-    #ifdef WINAPI_FAMILY
+    #if defined(WINAPI_FAMILY) && defined(WINAPI_FAMILY_PARTITION)
         #if WINAPI_FAMILY_PARTITION(WINAPI_PARTITION_APP)
             #define PPSSPP_PLATFORM_UWP 1
         #endif
